requirements.txt
# Core Dependencies
# PyTorch (CUDA 12.8) - Install with: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu128
torch==2.9.0
torchvision==0.24.0
torchaudio==2.9.0
triton==3.5.0; sys_platform == "linux"  # Linux only

# AI/ML Framework
accelerate==1.12.0
transformers==4.57.3
diffusers==0.36.0
timm==1.0.22
peft==0.18.0
safetensors==0.7.0
sentencepiece==0.2.1
tokenizers==0.22.1

# LLM Inference
vllm==0.13.0
xgrammar==0.1.27
llguidance==1.3.0
outlines_core==0.2.11
lm-format-enforcer==0.11.3
flashinfer-python==0.5.3

# LLM APIs
anthropic==0.71.0
openai==2.14.0
openai-harmony==0.0.8

# NVIDIA/CUDA (Linux/Windows with CUDA)
nvidia-ml-py==13.590.44
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.27.5
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.17.0
nvidia-cutlass-dsl==4.3.4
nvidia-nvshmem-cu12==3.3.20

# CUDA Python
cupy-cuda12x==13.6.0
cuda-python==13.1.1
cuda-bindings==13.1.1
cuda-pathfinder==1.3.3
fastrlock==0.8.3

# Deep Learning
einops==0.8.1
kornia==0.8.2
kornia_rs==0.1.10; sys_platform == "linux"  # Platform-specific
bitsandbytes==0.49.0; sys_platform == "linux"  # Linux only
compressed-tensors==0.12.2

# Computer Vision
opencv-python-headless==4.12.0.88  # Use opencv-python for desktop with GUI
pillow==12.0.0
imageio==2.37.2
scikit-image==0.26.0
rembg==2.0.69
PyMatting==1.1.14
tifffile==2025.12.20
av==16.0.1
qwen-vl-utils==0.0.14

# Web Framework
FastAPI==0.128.0
uvicorn==0.40.0
starlette==0.50.0
python-multipart==0.0.21
sse-starlette==3.1.1
gradio==6.2.0
gradio_client==2.0.2
fastapi-cli==0.0.20
fastapi-cloud-cli==0.8.0
uvloop==0.22.1
httptools==0.7.1
websockets==15.0.1

# Distributed Computing
ray==2.53.0
supervisor==4.3.0
setproctitle==1.3.7

# HTTP/Network
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
anyio==4.12.0
sniffio==1.3.1
websocket-client==1.9.0
httpx-sse==0.4.3
safehttpx==0.1.7
requests==2.32.5
urllib3==2.6.2
aiohttp==3.13.2
aiosignal==1.4.0
aiohappyeyeballs==2.6.1
frozenlist==1.8.0
multidict==6.7.0
yarl==1.22.0
propcache==0.4.1

# Data Science
numpy==2.2.6
pandas==2.3.3
matplotlib==3.10.8
scipy==1.16.3
networkx==3.6.1
sympy==1.14.0
numba==0.61.2
llvmlite==0.44.0

# Jupyter
jupyterlab>=4.5.0
jupyter_server>=2.17.0
jupyter_client>=8.7.0
ipykernel>=7.1.0
ipython>=9.8.0
ipywidgets>=8.1.0
jupyter-console>=6.6.0
notebook>=7.5.0
nbclient>=0.10.0
nbconvert>=7.16.0
nbformat>=5.10.0
notebook_shim>=0.2.0

# Utilities
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydantic-extra-types==2.11.0
python-dotenv==1.2.1
click==8.3.1
typer==0.21.0
rich==14.2.0
rich-toolkit==0.17.1
tqdm==4.67.1
coloredlogs==15.0.1
humanfriendly==10.0
tabulate==0.9.0
loguru==0.7.3

# File Processing
aiofiles==24.1.0
fsspec==2025.12.0
pooch==1.8.2
pydub==0.25.1
ffmpy==1.0.0
ijson==3.4.0.post0
orjson==3.11.5

# Security & Auth
cryptography==46.0.3
PyJWT==2.10.1
cffi==2.0.0
pycparser==2.22

# ML/AI Utilities
huggingface-hub==0.36.0
hf-xet==1.2.0
onnxruntime==1.23.2
protobuf==6.33.2
flatbuffers==25.12.19
gguf==0.17.1
tiktoken==0.12.0
mistral_common==1.8.8
lark==1.2.2
interegular==0.3.3

# MCP (Model Context Protocol)
mcp==1.25.0
blake3==1.0.8
cbor2==5.8.0
jmespath==1.0.1
model-hosting-container-standards==0.1.12

# Testing
pytest>=8.4.0
pytest-asyncio==0.25.2
pytest-cov==7.0.0
pytest-html==4.1.1
pytest-metadata==3.1.1
coverage==7.13.1

# Development
helper-dev-utils==0.5.0
helper-plot-hangul==0.5.6
annotated-doc==0.0.4
docstring_parser==0.17.0

# Monitoring & Observability
prometheus-fastapi-instrumentator==7.1.0
sentry-sdk==2.48.0
py-cpuinfo==9.0.0

# Other
ftfy==6.3.2
regex==2025.11.3
semantic-version==2.10.0
shellingham==1.5.4
tomlkit==0.13.3
watchfiles==1.1.1
email-validator==2.3.0
dnspython==2.8.0
pycountry==24.6.1
diskcache==5.6.3
dill==0.4.0
cloudpickle==3.1.2
filelock==3.20.1
cachetools==6.2.4
partial-json-parser==0.2.1.1.post7
typing-inspection==0.4.2
pybase64==1.4.3
groovy==0.1.2
fastar==0.8.0
rignore==0.7.6
apache-tvm-ffi==0.1.7
astor==0.8.1
depyf==0.20.0
msgspec==0.20.0
msgpack==1.1.2
ninja==1.13.0
